+++ src/Executor/BasicJobExecutor.php
@@ -6,14 +6,16 @@ use Closure;
 use Cron\CronExpression;
 use DateTimeImmutable;
 use Generator;
+use Orisai\Clock\Clock;
 use Orisai\Scheduler\Exception\RunFailure;
 use Orisai\Scheduler\Job\Job;
 use Orisai\Scheduler\Manager\JobManager;
 use Orisai\Scheduler\Status\JobSummary;
 use Orisai\Scheduler\Status\RunSummary;
-use Psr\Clock\ClockInterface;
 use Throwable;
+use function array_keys;
 use function assert;
+use function max;
 
 /**
  * @internal
@@ -21,17 +23,17 @@ use function assert;
 final class BasicJobExecutor implements JobExecutor
 {
 
-	private ClockInterface $clock;
+	private Clock $clock;
 
 	private JobManager $jobManager;
 
-	/** @var Closure(string|int, Job, CronExpression): array{JobSummary, Throwable|null} */
+	/** @var Closure(string|int, Job, CronExpression, int<0, max>): array{JobSummary, Throwable|null} */
 	private Closure $runCb;
 
 	/**
-	 * @param Closure(string|int, Job, CronExpression): array{JobSummary, Throwable|null} $runCb
+	 * @param Closure(string|int, Job, CronExpression, int<0, max>): array{JobSummary, Throwable|null} $runCb
 	 */
-	public function __construct(ClockInterface $clock, JobManager $jobManager, Closure $runCb)
+	public function __construct(Clock $clock, JobManager $jobManager, Closure $runCb)
 	{
 		$this->clock = $clock;
 		$this->jobManager = $jobManager;
@@ -40,29 +42,95 @@ final class BasicJobExecutor implements JobExecutor
 
 	public function runJobs(array $ids, DateTimeImmutable $runStart): Generator
 	{
+		if ($ids === []) {
+			return new RunSummary($runStart, $runStart, []);
+		}
+
+		$scheduledJobsBySecond = $this->getScheduledJobsBySecond($ids);
+		$lastSecond = max(array_keys($scheduledJobsBySecond));
+
 		$jobSummaries = [];
-		$suppressed = [];
-		foreach ($ids as $id) {
-			$scheduledJob = $this->jobManager->getScheduledJob($id);
-			assert($scheduledJob !== null);
-			[$job, $expression] = $scheduledJob;
+		$suppressedExceptions = [];
+		for ($second = 0; $second <= $lastSecond; $second++) {
+			$secondInitiatedAt = $this->clock->now();
 
-			[$jobSummary, $throwable] = ($this->runCb)($id, $job, $expression);
+			foreach ($scheduledJobsBySecond[$second] ?? [] as [$id, $job, $expression]) {
+				[$jobSummary, $throwable] = ($this->runCb)($id, $job, $expression, $second);
 
-			yield $jobSummaries[] = $jobSummary;
+				yield $jobSummaries[] = $jobSummary;
 
-			if ($throwable !== null) {
-				$suppressed[] = $throwable;
+				if ($throwable !== null) {
+					$suppressedExceptions[] = $throwable;
+				}
 			}
+
+			$this->sleepTillNextSecond($second, $lastSecond, $secondInitiatedAt);
 		}
 
 		$summary = new RunSummary($runStart, $this->clock->now(), $jobSummaries);
 
-		if ($suppressed !== []) {
-			throw RunFailure::create($summary, $suppressed);
+		if ($suppressedExceptions !== []) {
+			throw RunFailure::create($summary, $suppressedExceptions);
 		}
 
 		return $summary;
+	}
+
+	/**
+	 * @param non-empty-list<int|string> $ids
+	 * @return non-empty-array<int, list<array{int|string, Job, CronExpression}>>
+	 */
+	private function getScheduledJobsBySecond(array $ids): array
+	{
+		$scheduledJobsBySecond = [];
+		foreach ($ids as $id) {
+			$scheduledJob = $this->jobManager->getScheduledJob($id);
+			assert($scheduledJob !== null);
+			[$job, $expression, $repeatAfterSeconds] = $scheduledJob;
+
+			if ($repeatAfterSeconds === 0) {
+				$scheduledJobsBySecond[0][] = [$id, $job, $expression];
+			} else {
+				for ($second = 0; $second <= 59; $second += $repeatAfterSeconds) {
+					$scheduledJobsBySecond[$second][] = [$id, $job, $expression];
+				}
+			}
+		}
+
+		// $ids are not empty and for cycle is always run at least once
+		assert($scheduledJobsBySecond !== []);
+
+		return $scheduledJobsBySecond;
+	}
+
+	/**
+	 * More accurate than (float) $dateTime->format('U.u')
+	 */
+	private function getMicroTimestamp(DateTimeImmutable $dateTime): float
+	{
+		$seconds = (float) $dateTime->format('U');
+		$microseconds = (float) $dateTime->format('u') / 1e6;
+
+		return $seconds + $microseconds;
+	}
+
+	private function sleepTillNextSecond(int $second, int $lastSecond, DateTimeImmutable $secondInitiatedAt): void
+	{
+		$sleepTime = $this->getTimeTillNextSecond($second, $lastSecond, $secondInitiatedAt);
+		$this->clock->sleep(0, 0, (int) ($sleepTime * 1e6));
+	}
+
+	private function getTimeTillNextSecond(int $second, int $lastSecond, DateTimeImmutable $secondInitiatedAt): float
+	{
+		if ($second === $lastSecond) {
+			return 0;
+		}
+
+		$startOfSecond = $this->getMicroTimestamp($secondInitiatedAt);
+		$endOfSecond = $this->getMicroTimestamp($this->clock->now());
+		$timeElapsed = $endOfSecond - $startOfSecond;
+
+		return 1 - $timeElapsed;
 	}
 
 }