+++ src/Scheduler.php
@@ -8,13 +8,14 @@ use Orisai\Scheduler\Exception\JobFailure;
 use Orisai\Scheduler\Exception\RunFailure;
 use Orisai\Scheduler\Job\Job;
 use Orisai\Scheduler\Status\JobSummary;
+use Orisai\Scheduler\Status\RunParameters;
 use Orisai\Scheduler\Status\RunSummary;
 
 interface Scheduler
 {
 
 	/**
-	 * @return array<int|string, array{Job, CronExpression}>
+	 * @return array<int|string, array{Job, CronExpression, int<0, 30>}>
 	 */
 	public function getScheduledJobs(): array;
 
@@ -35,6 +36,6 @@ interface Scheduler
 	 * @phpstan-return ($force is true ? JobSummary : JobSummary|null)
 	 * @throws JobFailure When job failed and no error handler was set
 	 */
-	public function runJob($id, bool $force = true): ?JobSummary;
+	public function runJob($id, bool $force = true, ?RunParameters $parameters = null): ?JobSummary;
 
 }