+++ src/Status/JobInfo.php
@@ -14,16 +14,21 @@ final class JobInfo
 
 	private string $expression;
 
+	/** @var int<0, max> */
+	private int $second;
+
 	private DateTimeImmutable $start;
 
 	/**
 	 * @param string|int $id
+	 * @param int<0, max> $second
 	 */
-	public function __construct($id, string $name, string $expression, DateTimeImmutable $start)
+	public function __construct($id, string $name, string $expression, int $second, DateTimeImmutable $start)
 	{
 		$this->id = $id;
 		$this->name = $name;
 		$this->expression = $expression;
+		$this->second = $second;
 		$this->start = $start;
 	}
 
@@ -43,6 +48,14 @@ final class JobInfo
 	public function getExpression(): string
 	{
 		return $this->expression;
+	}
+
+	/**
+	 * @return int<0, max>
+	 */
+	public function getSecond(): int
+	{
+		return $this->second;
 	}
 
 	public function getStart(): DateTimeImmutable